requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1
cryptography==40.0.2
Werkzeug==2.3.7
python-dotenv==1.0.0
